ml/rembg-service/requirements.txt
flask>=3.0
rembg>=2.0
pillow>=10.0
